partial match notification part 2 and e2e test
